v1.40.10 Released
Fixed Bitfinex trade history bug
Fixed Bitstamp certificate issue
Switched to OpenSSL 1.1
Dropped support of Windows XP and macOS 10.9, but you still can compile
it manually.
Minor improvements

+++ src/exchange/exchange_bitfinex.h
@@ -1,6 +1,6 @@
 //  This file is part of Qt Bitcoin Trader
 //      https://github.com/JulyIGHOR/QtBitcoinTrader
-//  Copyright (C) 2013-2017 July IGHOR <julyighor@gmail.com>
+//  Copyright (C) 2013-2018 July IGHOR <julyighor@gmail.com>
 //
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
@@ -58,7 +58,7 @@ private:
 
     quint64 lastTradesDate;
     quint32 tickerLastDate;
-    quint32 lastHistoryId;
+    quint64 lastHistoryId;
 
     QList<DepthItem>* depthAsks;
     QList<DepthItem>* depthBids;
